+++ Amazing-Products/app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name="com.cs4520.assignment1.MainActivity"
+            android:name="com.cs4520.assignment1.activities.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.Assignment1">
             <intent-filter>
