+++ Amazing-Products/app/src/main/java/com/cs4520/assignment1/fragments/ProductListFragment.kt
@@ -1,7 +1,8 @@
-package com.cs4520.assignment1
+package com.cs4520.assignment1.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
+import com.cs4520.assignment1.R
 
 class ProductListFragment : Fragment(R.layout.product_list_fragment) {
     override fun onCreate(savedInstanceState: Bundle?) {
