+++ Amazing-Products/app/src/main/res/layout/product_list_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProductListFragment">
+    tools:context=".fragments.ProductListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/product_list"
