+++ Amazing-Products/app/src/main/res/values/dimens.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <dimen name="list_item_height">100dp</dimen>
+    <dimen name="list_image_size">80dp</dimen>
 </resources>
