+++ Amazing-Products/app/src/main/res/values/strings.xml
@@ -5,4 +5,5 @@
     <string name="username">Username</string>
     <string name="password">Password</string>
     <string name="login">Login</string>
+    <string name="food">food</string>
 </resources>
